Replaced old workaround for titles with a more native approach with Next.js
