+++ src/app/[lang]/(main)/about/page.tsx
@@ -16,21 +16,39 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 **/
 
-'use client';
-
 import Link from 'next/link';
 import Image from 'next/image';
 
 import style from './page.module.css';
-import { useDictionary } from '@/components/providers/DictionaryProvider';
 
 import profilePicture from '@/assets/img/profile-picture.webp';
 import ExternalLink from '@/components/icons/ExternalLink';
+import { getDictionary } from '@/app/dictionaries';
+import { Metadata } from 'next';
 
 
-export default function Page()
+export async function generateMetadata({
+	params,
+}: {
+	params: Promise<{ lang: Localizations }>,
+}): Promise<Metadata>
 {
-	const dict = useDictionary();
+	const lang = (await params).lang;
+	const dict = await getDictionary(lang);
+
+	return {
+		title: dict.pages.about.h1,
+	};
+}
+
+
+export default async function Page({
+	params,
+}: {
+	params: Promise<{ lang: Localizations }>,
+})
+{
+	const dict = await getDictionary((await params).lang);
 
 	return (
 		<main className={ style.main }>
