+++ src/app/[lang]/(main)/contact/page.tsx
@@ -16,23 +16,40 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 **/
 
-'use client';
-
 import Link from 'next/link';
+import { Metadata } from 'next';
 
 import style from './page.module.css';
+import { getDictionary } from '@/app/dictionaries';
 
 import Messenger from '@/components/icons/social/Messenger';
 import Telegram from '@/components/icons/social/Telegram/intex';
 import MailIcon from '@/components/icons/MailIcon';
 import LinkedIn from '@/components/icons/social/LinkedIn';
 import DiscordButton from './discord-button';
-import { useDictionary } from '@/components/providers/DictionaryProvider';
 
 
-export default function Page()
+export async function generateMetadata({
+	params,
+}: {
+	params: Promise<{ lang: Localizations }>,
+}): Promise<Metadata>
 {
-	const dict = useDictionary();
+	const lang = (await params).lang;
+	const dict = await getDictionary(lang);
+
+	return {
+		title: dict.pages.contact.h1,
+	};
+}
+
+export default async function Page({
+	params,
+}: {
+	params: Promise<{ lang: Localizations }>,
+})
+{
+	const dict = await getDictionary((await params).lang);
 
 	return (
 		<main className={ style.main }>
